# swagger.yaml
openapi: 3.0.3
info:
  title: StrikeMetrics RAG API
  description: >
    API para consultas RAG sobre estatísticas de combate e bomba do CS:GO.
    Utiliza Milvus como vector store e modelos Ollama locais para geração de respostas.
    Sprint 7 — AC2.
  version: 1.0.0

servers:
  - url: http://localhost:8000
    description: Servidor local de desenvolvimento

tags:
  - name: health
    description: Verificação de saúde da API
  - name: config
    description: Configurações e metadados da API
  - name: rag
    description: Pipeline RAG — Recuperação e Geração
  - name: admin
    description: Operações administrativas (uso com cuidado)

paths:

  /:
    get:
      tags: [health]
      summary: Health Check
      operationId: healthCheck
      responses:
        "200":
          description: API funcionando normalmente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
                  service: { type: string, example: StrikeMetrics RAG API }

  /models:
    get:
      tags: [config]
      summary: Listar Modelos Disponíveis
      operationId: listModels
      responses:
        "200":
          description: Lista de modelos retornada com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelsResponse"

  /metadata:
    get:
      tags: [config]
      summary: Metadados da API
      description: >
        Retorna informações sobre a API: versão, modelos suportados,
        vector store, embedding model, fontes de dados, endpoints e timestamp.
      operationId: getMetadata
      responses:
        "200":
          description: Metadados retornados com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetadataResponse"
              example:
                api_version: "1.0.0"
                supported_models: [llama3.2, mistral]
                vector_store: Milvus
                embedding_model: nomic-embed-text (Ollama)
                data_sources:
                  - csgo_combat_stats (Kaggle → MinIO Gold)
                  - csgo_bomb_stats (Kaggle → MinIO Gold)
                endpoints: [/, /models, /metadata, /query, /rebuild-index]
                timestamp: "2025-01-15T14:30:00Z"

  /query:
    post:
      tags: [rag]
      summary: Executar Consulta RAG
      description: >
        Executa o pipeline RAG completo (recuperação → contextualização → geração).

        **Validações de input:**
          - `model_name` deve estar na lista de GET /models
          - `question` entre 5 e 500 caracteres, com ao menos uma letra

        **Validações de output:**
          - Resposta gerada não pode ser vazia
      operationId: runQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
            examples:
              combate:
                summary: Dano de arma
                value:
                  model_name: llama3.2
                  question: Qual arma causa mais dano médio na cabeça?
              bomba:
                summary: Pós-plant
                value:
                  model_name: mistral
                  question: Qual site da bomba tem mais dano pós-plant?
      responses:
        "200":
          description: Resposta gerada com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
        "422":
          description: Erro de validação de input (model_name inválido, question fora do tamanho, etc).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
              examples:
                modelo_invalido:
                  summary: Modelo não suportado
                  value:
                    detail:
                      - loc: [body, model_name]
                        msg: "Modelo 'gpt-4' não suportado."
                        type: value_error
                pergunta_curta:
                  summary: Pergunta muito curta
                  value:
                    detail:
                      - loc: [body, question]
                        msg: "String should have at least 5 characters"
                        type: string_too_short

  /rebuild-index:
    post:
      tags: [admin]
      summary: Reconstruir Índice Vetorial
      description: >
        Apaga e recria toda a coleção no Milvus a partir dos dados do MinIO Gold.
        Use com cuidado.
      operationId: rebuildIndex
      responses:
        "200":
          description: Índice reconstruído com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RebuildResponse"
        "500":
          description: Erro interno.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"

components:
  schemas:

    QueryRequest:
      type: object
      required: [question]
      properties:
        model_name:
          type: string
          default: llama3.2
          description: Modelo Ollama. Deve ser um valor retornado por GET /models.
          example: llama3.2
        question:
          type: string
          minLength: 5
          maxLength: 500
          description: Pergunta sobre estatísticas de CS:GO (5–500 chars, ao menos uma letra).
          example: Qual arma causa mais dano médio na cabeça?

    QueryResponse:
      type: object
      required: [answer, logs]
      properties:
        answer:
          type: string
          minLength: 1
          description: Resposta gerada pelo LLM. Nunca será vazia.
        logs:
          type: string
          description: Logs internos do pipeline RAG.

    MetadataResponse:
      type: object
      required: [api_version, supported_models, vector_store, embedding_model, data_sources, endpoints, timestamp]
      properties:
        api_version:
          type: string
          description: Versão atual da API.
        supported_models:
          type: array
          items: { type: string }
          description: Modelos Ollama disponíveis.
        vector_store:
          type: string
          description: Tecnologia de vector store utilizada.
        embedding_model:
          type: string
          description: Modelo de embedding utilizado para indexação.
        data_sources:
          type: array
          items: { type: string }
          description: Fontes de dados indexadas.
        endpoints:
          type: array
          items: { type: string }
          description: Endpoints disponíveis na API.
        timestamp:
          type: string
          format: date-time
          description: Data/hora da consulta (ISO 8601 UTC).

    ModelsResponse:
      type: object
      required: [models]
      properties:
        models:
          type: array
          items: { type: string }

    RebuildResponse:
      type: object
      properties:
        message: { type: string }

    HTTPError:
      type: object
      properties:
        detail: { type: string }

    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items: { type: string }
              msg: { type: string }
              type: { type: string }